Fixes for clean NX compile/link
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@2374 42af7a65-404d-4744-a932-0658087f49c3

+++ graphics/nxfonts/nxfonts_convert.c
@@ -40,6 +40,7 @@
 #include <nuttx/config.h>
 
 #include <stdint.h>
+#include <stddef.h>
 #include <debug.h>
 
 #include <nuttx/nxglib.h>
@@ -145,7 +146,7 @@ int NXF_FUNCNAME(nxf_convert,NXFONTS_SUFFIX)
   FAR NXF_PIXEL_T *dptr;
   FAR const uint8_t *sptr;
   uint8_t bmbyte;
-  uint8_t bmbit;
+  int bmbit;
   int row;
   int col;
   int bmndx;
